requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
flask-migrate
psycopg2-binary==2.9.9
bcrypt==4.1.2
Pillow==10.2.0
python-dotenv==1.0.0
gunicorn==21.2.0
python-socketio==5.11.1
eventlet==0.35.2
numpy>=1.24.0
opencv-python-headless>=4.8.0
google-genai>=0.2.0
google-generativeai
openpyxl>=3.1.0
pandas>=2.0.0
xlrd>=2.0.1
sqlalchemy
requests
gevent
gevent-websocket
python-dateutil
werkzeug
bcrypt
eventlet
Flask
Flask-CORS
Flask-JWT-Extended
flask-migrate
Flask-SocketIO
Flask-SQLAlchemy
gevent
gevent-websocket
google-genai
google-generativeai
gunicorn
numpy
opencv-python-headless
openpyxl
pandas
Pillow
psycopg2-binary
python-dateutil
python-dotenv
python-socketio
requests
sqlalchemy
werkzeug
xlrd
bcrypt
eventlet
Flask
Flask-CORS
Flask-JWT-Extended
flask-migrate
Flask-SocketIO
Flask-SQLAlchemy
gevent
gevent-websocket
google-genai
google-generativeai
gunicorn
numpy
opencv-python-headless
openpyxl
pandas
Pillow
psycopg2-binary
python-dateutil
python-dotenv
python-socketio
requests
sqlalchemy
werkzeug
xlrd
bcrypt
eventlet
flask
flask-cors
flask-jwt-extended
flask-migrate
flask-socketio
flask-sqlalchemy
gevent
gevent-websocket
google-genai
google-generativeai
gunicorn
numpy
opencv-python-headless
openpyxl
pandas
pillow
psycopg2-binary
python-dateutil
python-dotenv
python-socketio
requests
sqlalchemy
werkzeug
xlrd
bcrypt
eventlet
flask
flask-cors
flask-jwt-extended
flask-migrate
flask-socketio
flask-sqlalchemy
gevent
gevent-websocket
google-genai
google-generativeai
gunicorn
numpy
opencv-python-headless
openpyxl
pandas
pillow
psycopg2-binary
python-dateutil
python-dotenv
python-socketio
requests
sqlalchemy
werkzeug
xlrd
bcrypt
eventlet
flask
flask-cors
flask-jwt-extended
flask-migrate
flask-socketio
flask-sqlalchemy
gevent
gevent-websocket
google-genai
google-generativeai
gunicorn
numpy
opencv-python-headless
openpyxl
pandas
pillow
psycopg2-binary
python-dateutil
python-dotenv
python-socketio
requests
sqlalchemy
werkzeug
xlrd
bcrypt
eventlet
flask
flask-cors
flask-jwt-extended
flask-migrate
flask-socketio
flask-sqlalchemy
gevent
gevent-websocket
google-genai
google-generativeai
gunicorn
numpy
opencv-python-headless
openpyxl
pandas
pillow
psycopg2-binary
python-dateutil
python-dotenv
python-socketio
requests
sqlalchemy
werkzeug
xlrd
